requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
feedparser==6.0.10

# Storage
supabase==2.3.0

# AI/Translation
google-generativeai==0.3.2

# WordPress
requests==2.31.0

